+++ otter/assign/output.py
@@ -70,7 +70,7 @@ def write_autograder_dir(nb_path, output_nb_path, assignment, args):
             # check that file is in subdir
             assert os.path.abspath(nb_path.parent) in os.path.abspath(file), \
                 f"{file} is not in a subdirectory of the master notebook directory"
-            file_path = pathlib.Path(file)
+            file_path = pathlib.Path(file).resolve()
             rel_path = file_path.parent.relative_to(nb_path.parent)
             os.makedirs(output_dir / rel_path, exist_ok=True)
             shutil.copy(file, str(output_dir / rel_path))